% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDiscretization.R
\name{AnyObsByLine}
\alias{AnyObsByLine}
\title{function that return 1 for a vector without observation after/during discretization}
\usage{
AnyObsByLine(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a binary
}
\description{
function that return 1 for a vector without observation after/during discretization
}
