% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CInLPN.formula.R
\name{CInLPN}
\alias{CInLPN}
\title{Causal Inference in a Latent Processes Network}
\usage{
CInLPN(structural.model, measurement.model, parameters, option, Time,
  subject, data, ...)
}
\arguments{
\item{structural.model}{a list of arguments used to specify the structural model
The structural model contains  five component to specify it.}

\item{measurement.model}{a list of arguments used to specify the measurement model}

\item{parameters}{a list of arguments about parameters of the models (e.g., initial paremeters, parameters one would like to fix, etc.)}

\item{option}{a list of arguments for optimization purpose}

\item{Time}{name of the covariate representing the time}

\item{subject}{name of the covariate representing the grouping structure}

\item{data}{data frame containing the variables named in strutural model, time, subject.}

\item{structural.model$fixed.LP0}{allow to specify a fixed effects model on the baseline level of processes
Note that there is no need to specify random effect model for the baseline level of processes. As we
set systematicaly an random intercept with variance 1 for identifiability purpose}

\item{structural.model$fixed.DeltaLP}{a two-sided linear formula object for specifying the fixed-effects in the linear mixed 
model for initial level of latent processes. The response outcome is on the left of ~ symbol and the covariates are separated by + on the 
right of the ~ symbol. Fo identifiability purposes, the intercept of model are fixed to 0 (not estimated).}

\item{structural.model$random.DeltaLP}{allow to specify a random effects model on the processes slopes}

\item{structural.model$trans.matrix}{allow to specify a model for element of the transition matrix, which capture 
temporal influences between latent processes.}

\item{structural.model$delta.time}{to indicates the discretisation step to be used for latent processes}

\item{measurement.model$link}{to indicate link to be used to transform markers. Takes values in ("linear", "n-type-d")
where "n" indicates the number of nodes, "type" : takes values in ("quant", "manual", "equi") to indicates how nodes are placed, 
and finaly argument "d" indicates the degre of I-splilnes to be used to construct the link function}

\item{measurement.model$knots}{argument to indicate if necessary the place of kodes, default value is NULL}

\item{parameters$paras.ini}{to indicates initial values for parameters, default values is NULL}

\item{parameters$Fixed.para.indix}{to indicates the position of parameters to be constraint.}

\item{parameters$Fixed.para.values}{to indicates the values associates to the index of parameters to be constraint.}

\item{option$epsa}{convergence criteria}

\item{option$epsb}{convergence criteria}

\item{option$epsc}{convergence criteria}

\item{option$MCnr}{number of replik  to compute the predictions in real scales}
}
\value{
0
}
\description{
This package  estimates trajectories of multivariate latent processes observed through K longitudinal markers (K ≥ D) and
relationships  between latent processes. 
A structural model for processes network is defined for latent processes. This structural model combines 
multivariate linear mixed model and system of difference equations to model trajectories of the processes and 
their temporal influences other time. The temporal influences are captured through a time-depend transition matrix that can be 
adjusted for covariates. Instead of temporal influences, the variance-covariance matrix of random effects on processes can be suitably used to
modeling  (residual) correlations between latent processes.
Longitudinal markers are related to their latent processes through link function, which parameters are estimated simultaneously 
with  those of structural model. The link function can be linear for Gaussien marker or non linear (I-splines basis) for
non Gaussian markers.
All parameters are  estimated simultaneously in a maximum likelihood framework 
See the methodological paper available at :http://arxiv.org/abs/1806.03659
}
\examples{
### example 1
Delta <- 1
paras.ini <- NULL
indexparaFixeUser <- c(1,4,10+c(1,2,4,5,6,9, 10+c(1:4)))
paraFixeUser <- c(0,0,1,0,0,1,0,0, rep(0,4))
mod <-  CInLPN(structural.model = list(fixed.LP0 = ~ 1 + C1 + C2|1 + C1 + C2,
                                   fixed.DeltaLP = L1 + L2 | L3 ~1 + time|1 + time,
                                   random.DeltaLP = ~ 1|1,
                                   trans.matrix = ~ 1,
                                   delta.time = Delta),
           measurement.model = list(link.functions = list(links = c("4-equi-2", "linear", "4-equi-2"),
                                                          knots = list(NULL, NULL, NULL))),
           parameters = list(paras.ini = paras.ini, Fixed.para.index = indexparaFixeUser, Fixed.para.values = paraFixeUser),
           option = list(nproc = 1, print.info = F, mekepred = T, MCnr = 10, univarmaxiter = 7, epsa = 1e-5, epsb = 1e-5, epsd = 1e-5),
           Time = "time",
           subject = "id",
           data = data
         )
         
### example 2
library(splines)
Delta <- 0.5
paras.ini <- NULL
indexparaFixeUser <- c(1,3, 8+c(1, 2, 4, 5, 6, 9,10+c(2:4,14:16)))
paraFixeUser <- c(0, 0, 1, 0, 0, 1, 0, 0, rep(0,6))
res <- CInLPN(structural.model = list(fixed.LP0 = ~ 1 + C2 | 1 + C2,
                                 fixed.DeltaLP = L1 | L2  ~ 1 + time| 1 + time ,
                                 random.DeltaLP = ~ 1|1,
                                 trans.matrix = ~ 1 + bs(x = time, knots =c(2), intercept = F, degree = 2),
                                 delta.time = Delta),
         measurement.model = list(link.functions = list(links = c(NULL,NULL),
                                                        knots = list(NULL, NULL))),
         
         parameters = list(paras.ini = paras.ini, Fixed.para.index = indexparaFixeUser, Fixed.para.values = paraFixeUser),
         option = list(nproc = 2, print.info = T, mekepred = T, MCnr = 10, univarmaxiter = 7, epsa = 1e-5, epsb = 1e-4, epsd = 1e-2),
         Time = "time",
         subject = "id",
         data = data
  )
    
}
