% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compoYiNA}
\alias{compoYiNA}
\title{After vectorising the vzector Yi, this function returns
a vector indicating missing values : 1 = observed value, 0 = missing value}
\usage{
compoYiNA(Yi)
}
\arguments{
\item{Yi}{a matrix with possibly NAs}
}
\value{
a vector of elements 0,1
}
\description{
After vectorising the vzector Yi, this function returns
a vector indicating missing values : 1 = observed value, 0 = missing value
}
