% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ConstrA}
\alias{ConstrA}
\title{Function that  constructs the transition matrix at time t}
\usage{
ConstrA(K, t, DeltaT, vec_alpha_ij, modA_mat)
}
\arguments{
\item{K}{an integer representing the size of K*K matrix}

\item{t}{an integer indicating the time at which coefficients are computed}

\item{DeltaT}{double that indicates the discretization step}

\item{vec_alpha_ij}{a vector of overall parameters associated to the
model.matrix for elements of the transistion matrix}

\item{modA_mat}{model.matrix for elements of the transistion matrix}
}
\value{
Id + DeltaT*A where A is the transition matrix
}
\description{
Function that  constructs the transition matrix at time t
}
