% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{simulated data}
\format{A dataset of 3584 repeated observations (for X independent subjects) and 7 variables
(id=subject identification number, C1= continuous time-independent covariate, C2=binary time-independent covariate,
time=observation time, L1-L2-L3=3 longitudinal markers)}
\usage{
data(data)
}
\description{
simulated data
}
\examples{
data(data)
head(data)
}
\keyword{datasets}
