% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat.R
\name{DataFormat}
\alias{DataFormat}
\title{Function to formated the data}
\usage{
DataFormat(data, subject, fixed_X0.models, randoms_X0.models,
  fixed_DeltaX.models, randoms_DeltaX.models, mod_trans.model,
  link = NULL, knots = NULL, outcomes, nD, Time, DeltaT)
}
\arguments{
\item{data}{indicates the data frame containing all the variables for estimating the model}

\item{subject}{indicates the name of the covariate representing the grouping structure}

\item{fixed_X0.models}{fixed effects in the submodel for the baseline level of processes}

\item{randoms_X0.models}{random effects in the submodel for the baseline level of processes}

\item{fixed_DeltaX.models}{a two-sided linear formula object for specifying the response outcomes (one the left part of ~ symbol) 
and the covariates with fixed-effects (on the right part of ~ symbol)}

\item{randoms_DeltaX.models}{random effects in the submodel for change over time of latent processes}

\item{mod_trans.model}{model for elements of the temporal transition matrix, which captures 
the temporal influences between latent processes}

\item{link}{indicates link used to transform outcome}

\item{knots}{indicates position of knots used to transform outcomes}

\item{outcomes}{names of the outcomes}

\item{nD}{number of latent processes}

\item{Time}{indicates the name of the covariate representing the time}

\item{DeltaT}{indicates the discretization step}
}
\value{
a list
}
\description{
Function to formated the data
}
