% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DparChol}
\alias{DparChol}
\title{Function that computes a symetric D matric from it Cholesky L}
\usage{
DparChol(q, qvector)
}
\arguments{
\item{q}{an integer}

\item{qvector}{a vector of length q}
}
\value{
a symetric positive define matrix
}
\description{
Function that computes a symetric D matric from it Cholesky L
}
