% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat.R
\name{f.link}
\alias{f.link}
\title{Generate model.matrix for transformation of outcomes}
\usage{
f.link(outcomes, Y, link = NULL, knots = NULL, na.action = "na.pass")
}
\arguments{
\item{outcomes}{names of the outcomes}

\item{Y}{values of the outcomes}

\item{link}{indicates link used to transform outcome}

\item{knots}{indicates position of knots used to transform outcomes}

\item{na.action}{indicates na.action argument}
}
\value{
a list
}
\description{
Generate model.matrix for transformation of outcomes
}
