% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mvrnorm.R
\name{f_mvrnorm}
\alias{f_mvrnorm}
\title{Generate a multivate normal vector}
\usage{
f_mvrnorm(seed, m, sd)
}
\arguments{
\item{seed}{seed for the randomization}

\item{m}{mean of distribution}

\item{sd}{standard deviation of the distribution}
}
\value{
a multivate normal vector
}
\description{
Generate a multivate normal vector
}
