% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametre.R
\name{f_paras.ini}
\alias{f_paras.ini}
\title{Initialisation of parameters}
\usage{
f_paras.ini(data, outcomes, mapped.to.LP, fixed_X0.models,
  fixed_DeltaX.models, randoms_DeltaX.models, randoms_X0.models, nb_RE,
  mod_trans.model, subject, Time, link, knots, DeltaT, maxiter = 25,
  epsa = 1e-04, epsb = 1e-04, epsd = 1e-04, nproc = 1,
  print.info = TRUE)
}
\arguments{
\item{data}{indicates the data frame containing all the variables for estimating the model}

\item{outcomes}{names of the outcomes}

\item{mapped.to.LP}{indicates which outcome measured which latent process, it is a mapping table between
outcomes and latents processes}

\item{fixed_X0.models}{fixed effects in the submodel for the baseline level of processes}

\item{fixed_DeltaX.models}{a two-sided linear formula object for specifying the response outcomes (one the left part of ~ symbol) 
and the covariates with fixed-effects (on the right part of ~ symbol)}

\item{randoms_DeltaX.models}{random effects in the submodel for change over time of latent processes}

\item{randoms_X0.models}{random effects in the submodel for the baseline level of processes}

\item{nb_RE}{number of random effects}

\item{mod_trans.model}{model for elements of the temporal transition matrix, which captures 
the temporal influences between latent processes}

\item{subject}{indicates the name of the covariate representing the grouping structure}

\item{Time}{indicates the name of the covariate representing the time}

\item{link}{indicates link used to transform outcome}

\item{knots}{indicates position of knots used to transform outcomes}

\item{DeltaT}{indicates the discretization step}

\item{maxiter}{maximum iteration}

\item{epsa}{threshold for the convergence criterion on the parameters, default value is 1.e-4}

\item{epsb}{threshold for the convergence criterion on the likelihood, default value is 1.e-4}

\item{epsd}{threshold for the convergence criterion on the derivatives, default value is 1.e-3}

\item{nproc}{number of processor to be used for running this package}

\item{print.info}{to print information during the liklihood optimization, default value is FALSE}
}
\value{
a list
}
\description{
Initialisation of parameters
}
