% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDiscretization.R
\name{f_TempsDiscr}
\alias{f_TempsDiscr}
\title{function that discretized a vector of time with a given delta.}
\usage{
f_TempsDiscr(Time, Delta)
}
\arguments{
\item{Time}{input continuous time vector}

\item{Delta}{discretized step}
}
\value{
a discretized  time vector
}
\description{
function that discretized a vector of time with a given delta.
}
