% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_trSpline.R
\name{f_trSpline}
\alias{f_trSpline}
\title{Function that compute  transformaitions  by I-splines basis and their derivates}
\usage{
f_trSpline(y, minY, maxY, knots, degree, paras)
}
\arguments{
\item{y}{outcomes in real scales}

\item{minY}{minimum of outcome y}

\item{maxY}{maximum of outcome y}

\item{knots}{indicates position of knots used to transform outcomes}

\item{degree}{indicates degree of the basis of splines}

\item{paras}{initial values for parameters}
}
\value{
a matrix
}
\description{
Function that compute  transformaitions  by I-splines basis and their derivates
}
