% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericFunctions.R
\name{f_var_cov}
\alias{f_var_cov}
\title{function to create the variance-covariance matrix  from vector of all element of the matrix}
\usage{
f_var_cov(model.b, model.v)
}
\arguments{
\item{model.b}{estimates of parameters}

\item{model.v}{variance-covariance elements}
}
\value{
a matrix
}
\description{
function to create the variance-covariance matrix  from vector of all element of the matrix
}
