% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_Yi_r_NA_by0}
\alias{f_Yi_r_NA_by0}
\title{Function that replaces NAs by 0.0 just for computatinal need}
\usage{
f_Yi_r_NA_by0(Yi)
}
\arguments{
\item{Yi}{a matrix with possibly NAs}
}
\value{
a matrix
}
\description{
Function that replaces NAs by 0.0 just for computatinal need
}
