% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GmatA0totaui}
\alias{GmatA0totaui}
\title{Function that  creates a matrix K,K*(max(tau_i)-1) containing  sub-matrices {(A_t)}_{t=0,tau_i-1}}
\usage{
GmatA0totaui(K, vec_alpha_ij, tau_i, DeltaT, modA_mat)
}
\arguments{
\item{K}{an integer representing the size of K*K matrix}

\item{vec_alpha_ij}{a vector of overall parameters associated to the
model.matrix for elements of the transistion matrix}

\item{tau_i}{vector of integers indicating times at which coefficients are computed}

\item{DeltaT}{double that indicates the discretization step}

\item{modA_mat}{model.matrix for elements of the transistion matrix}
}
\value{
a matrix containing matrix of form Id + DeltaT*A
}
\description{
Function that  creates a matrix K,K*(max(tau_i)-1) containing  sub-matrices {(A_t)}_{t=0,tau_i-1}
}
