% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GmatprodAstotau}
\alias{GmatprodAstotau}
\title{Function that  creates a big matrix containing  Prod(A_t)t=t_ini,tau.}
\usage{
GmatprodAstotau(K, vec_alpha_ij, tau, t_ini, DeltaT, modA_mat)
}
\arguments{
\item{K}{an integer representing the size of K*K matrix}

\item{vec_alpha_ij}{a vector of overall parameters associated to the
model.matrix for elements of the transistion matrix}

\item{tau}{vector of integers indicating times}

\item{t_ini}{indicates the started discretized time}

\item{DeltaT}{double that indicates the discretization step}

\item{modA_mat}{model.matrix for elements of the transistion matrix}
}
\value{
a matrix
}
\description{
Function that  creates a big matrix containing  Prod(A_t)t=t_ini,tau.
}
