% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericFunctions.R
\name{inclu_intercerpt}
\alias{inclu_intercerpt}
\title{Function used to replace 1 by intercept in a formula}
\usage{
inclu_intercerpt(x)
}
\arguments{
\item{x}{vector of text}
}
\value{
vector of characters
}
\description{
Function used to replace 1 by intercept in a formula
}
