% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat.R
\name{is_na_vec}
\alias{is_na_vec}
\title{function that return 1 if all element of a vector is NA}
\usage{
is_na_vec(vec)
}
\arguments{
\item{vec}{a numerical vector}
}
\value{
a vector
}
\description{
function that return 1 if all element of a vector is NA
}
