% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KmatDiag}
\alias{KmatDiag}
\title{Function that creates a K-block diagonal matrix with non diagonal element fixed to 0}
\usage{
KmatDiag(Kvector)
}
\arguments{
\item{Kvector}{a vector of length K}
}
\value{
a diagonal matrix
}
\description{
Function that creates a K-block diagonal matrix with non diagonal element fixed to 0
}
