% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Loglik}
\alias{Loglik}
\title{Function that computes the log-likelihood of the observed data}
\usage{
Loglik(K, nD, mapping, paraOpt, paraFixe, posfix, m_is, Mod_MatrixY,
  Mod_MatrixYprim, df, x, z, q, nb_paraD, x0, z0, q0, if_link, tau, tau_is,
  modA_mat, DeltaT)
}
\arguments{
\item{K}{an integer indicating the number of markers}

\item{nD}{an integer indicating the number of latent processes}

\item{mapping}{indicates which outcome measured which latent process, it is a mapping table between
outcomes and latents processes}

\item{paraOpt}{initial values for model parameters}

\item{paraFixe}{values associated to the index of parameters to be constrained}

\item{posfix}{position of parameters to be constrained}

\item{m_is}{vector of numbers of visit occasions for individuals}

\item{Mod_MatrixY}{model.matrix from markers transformation submodels}

\item{Mod_MatrixYprim}{model.matrix from the derivates of markers transformation submodels}

\item{df}{vector of numbers of parameters for each transformation model}

\item{x}{model.matrix for change's fixed submodel}

\item{z}{model.matrix for change's random effects submodel}

\item{q}{a vector of number of random effects on each change latent process over time}

\item{nb_paraD}{number of paramerters of the variance-covariance matrix of random effects}

\item{x0}{model.matrix for baseline's fixed submodel}

\item{z0}{model.matrix for baseline's random effects submodel}

\item{q0}{a vector of number of random effects on each initial latent process level}

\item{if_link}{indicates if non linear link is used to transform an outcome}

\item{tau}{a vector of integers indicating times (including maximum time)}

\item{tau_is}{a vector of integers indicating times for individuals}

\item{modA_mat}{model.matrix for elements of the transistion matrix}

\item{DeltaT}{double that indicates the discretization step}
}
\value{
double
}
\description{
Function that computes the log-likelihood of the observed data
}
