% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_trans.R
\name{mat_trans}
\alias{mat_trans}
\title{Function to create a matrix of all transition matrix and their standard deviation}
\usage{
mat_trans(model, K, data, ind, DeltaT, ind_parafixA, val_parafixA)
}
\arguments{
\item{model}{the model}

\item{K}{number of the outcomes}

\item{data}{indicates the data frame}

\item{ind}{indicates the order (rank) of the indivudal}

\item{DeltaT}{indicates the discretization step}

\item{ind_parafixA}{position of matrix A parameters to be constrained}

\item{val_parafixA}{values associated to the index of parameters of the matrix A to be constrained}
}
\value{
a list
}
\description{
Function to create a matrix of all transition matrix and their standard deviation
}
