% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matNui}
\alias{matNui}
\title{Function that constructs the matrix nu_t_j, the expectation of the processes at time t_j}
\usage{
matNui(nD, tau_i, DeltaT, x0i, alpha_mu0, xi, alpha_mu, G_mat_A_0_to_tau_i)
}
\arguments{
\item{nD}{an integer indicating the number of processes}

\item{tau_i}{vector of integers indicating times}

\item{DeltaT}{double that indicates the discretization step}

\item{x0i}{model.matrix for baseline's submodel}

\item{alpha_mu0}{a vector of parameters associated to the model.matrix for the baseline's submodel}

\item{xi}{model.matrix for change's submodel}

\item{alpha_mu}{a vector of parameters associated to the model.matrix for the change's submodel}

\item{G_mat_A_0_to_tau_i}{matrix containing  Prod(A_t)t=0,tau_i where A_t is the transition
matric containing at time t}
}
\value{
a matrix
}
\description{
Function that constructs the matrix nu_t_j, the expectation of the processes at time t_j
}
