% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDiscretization.R
\name{OneOutcomeAtLeast}
\alias{OneOutcomeAtLeast}
\title{function that filters initial data to return only individuals having at least one observation for outcomes onf interest.
Meaning that the returned data contained individuals with at least one measurement at the considered visit time.}
\usage{
OneOutcomeAtLeast(data, subject, outcomes)
}
\arguments{
\item{data}{input data}

\item{subject}{subject identifiant}

\item{outcomes}{a vector of outcomes of interest: markers}
}
\value{
a dataframe
}
\description{
function that filters initial data to return only individuals having at least one observation for outcomes onf interest.
Meaning that the returned data contained individuals with at least one measurement at the considered visit time.
}
