% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametre.R
\name{Parametre}
\alias{Parametre}
\title{Function to initialize parameters kin multivariate CInLPN model}
\usage{
Parametre(K, nD, vec_ncol_x0n, n_col_x, nb_RE, stochErr = FALSE,
  indexparaFixeUser = NULL, paraFixeUser = NULL, L = 1, paras.ini,
  ncolMod.MatrixY)
}
\arguments{
\item{K}{number of the markers}

\item{nD}{number of the latent processes}

\item{vec_ncol_x0n}{vector of number of columns of model.matrix for baseline's submodel}

\item{n_col_x}{number of overall columns of model.matrix for change's submodel}

\item{nb_RE}{number of random effects}

\item{stochErr}{indicates if the structural model contain stochastique components}

\item{indexparaFixeUser}{position of parameters to be constrained}

\item{paraFixeUser}{values associated to the index of parameters to be constrained}

\item{L}{number of columns of model.matrix for temporal infuences model}

\item{paras.ini}{initial values for parameters, default values is NULL}

\item{ncolMod.MatrixY}{vector of number of columns of model.matrix for transformation submodel}
}
\value{
a list
}
\description{
Function to initialize parameters kin multivariate CInLPN model
}
