% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_matA.R
\name{Plot_a}
\alias{Plot_a}
\title{Plot the temporal infmuences matrix A}
\usage{
Plot_a(recaps.A, with_IC = FALSE, Ylim = c(-2, 2), age0, legend,
  pos.legend, color, xlab = "Time (in year)")
}
\arguments{
\item{recaps.A}{values of matrix A}

\item{with_IC}{indicates if the plot is done with confidence band, default value is FALSE}

\item{Ylim}{minimum and maxmum of Y axe}

\item{age0}{indicates baseline age}

\item{legend}{indicates a legend}

\item{pos.legend}{indicates the legend position}

\item{color}{color to be used}

\item{xlab}{xlab}
}
\value{
---
}
\description{
Plot the temporal infmuences matrix A
}
