% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_X.R
\name{PlotX}
\alias{PlotX}
\title{Function to plot trajectories of observation and their predictions}
\usage{
PlotX(Obs, outcomes, preds, VarTime, with_IC = FALSE, Ylim,
  xlab = "Time", ylab = "Score", age0 = 0, Title)
}
\arguments{
\item{Obs}{Observations}

\item{outcomes}{names of outcomes}

\item{preds}{indicates predictions}

\item{VarTime}{indicates the name of the covariate representing the time}

\item{with_IC}{indicates if the plot is done with confidence band, default value is FALSE}

\item{Ylim}{minimum and maxmum of Y axe}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{age0}{indicates baseline age}

\item{Title}{Titles for the plots}
}
\value{
---
}
\description{
Function to plot trajectories of observation and their predictions
}
