% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred}
\alias{pred}
\title{Function that computes the predictions (marginal and subject-specific) for individuals}
\usage{
pred(K, nD, mapping, paras, m_is, Mod_MatrixY, df, x, z, q, nb_paraD, x0,
  z0, q0, if_link, tau, tau_is, modA_mat, DeltaT, MCnr, minY, maxY, knots,
  degree, epsPred)
}
\arguments{
\item{K}{an integer indicating the number of markers}

\item{nD}{an integer indicating the number of latent processes}

\item{mapping}{indicates which outcome measured which latent process, it is a mapping table between
outcomes and latents processes}

\item{paras}{values of model parameters}

\item{m_is}{vector of numbers of visit occasions for individuals}

\item{Mod_MatrixY}{model.matrix from markers transformation submodels}

\item{df}{vector of numbers of parameters for each transformation model}

\item{x}{model.matrix for change's fixed submodel}

\item{z}{model.matrix for change's random effects submodel}

\item{q}{a vector of number of random effects on each change latent process over time}

\item{nb_paraD}{number of paramerters of the variance-covariance matrix of random effects}

\item{x0}{model.matrix for baseline's fixed submodel}

\item{z0}{model.matrix for baseline's random effects submodel}

\item{q0}{a vector of number of random effects on each initial latent process level}

\item{if_link}{indicates if non linear link is used to transform an outcome}

\item{tau}{a vector of integers indicating times (including maximum time)}

\item{tau_is}{a vector of integers indicating times for individuals}

\item{modA_mat}{model.matrix for elements of the transistion matrix}

\item{DeltaT}{double that indicates the discretization step}

\item{MCnr}{an integer that indicates the number of sample for MC method}

\item{minY}{a vector of minima of outcomes}

\item{maxY}{a vector of maxima of outcomes}

\item{knots}{indicates position of knots used to transform outcomes}

\item{degree}{indicates degree of the basis of splines}

\item{epsPred}{convergence criteria for prediction using MC method}
}
\value{
a matrix
}
\description{
Function that computes the predictions (marginal and subject-specific) for individuals
}
