% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CInLPN.R
\name{predict.CInLPN}
\alias{predict.CInLPN}
\title{Title}
\usage{
\method{predict}{CInLPN}(model, newdata, MCnr = 10, ...)
}
\arguments{
\item{model}{CInLPN object}

\item{newdata}{dataset}

\item{MCnr}{an integer that give number of Monte Carlo iterations}
}
\value{
list of marginal and subject-specific predictions
}
\description{
Title
}
