% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CInLPN.R
\name{print.CInLPN}
\alias{print.CInLPN}
\title{Print CInLPN object}
\usage{
\method{print}{CInLPN}(x, ...)
}
\arguments{
\item{x}{an CInLPN object}

\item{\dots}{optional parameters}
}
\value{
0
}
\description{
Print CInLPN object
}
