% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ProdA}
\alias{ProdA}
\title{Function that  computes the product of A(t) for t1 to t2}
\usage{
ProdA(K, t2, t1, DeltaT, vec_alpha_ij, modA_mat)
}
\arguments{
\item{K}{an integer representing the size of K*K matrix}

\item{t2}{indicates the ended discretized time}

\item{t1}{indicates the started discretized time}

\item{DeltaT}{double that indicates the discretization step}

\item{vec_alpha_ij}{a vector of overall parameters associated to the
model.matrix for elements of the transistion matrix}

\item{modA_mat}{model.matrix for elements of the transistion matrix}
}
\value{
a matrix
}
\description{
Function that  computes the product of A(t) for t1 to t2
}
