% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CInLPN.R
\name{summary.CInLPN}
\alias{summary.CInLPN}
\title{Summary of CInLPN object}
\usage{
\method{summary}{CInLPN}(object, ...)
}
\arguments{
\item{object}{an CInLPN object}

\item{\dots}{optional parameters}
}
\value{
0
}
\description{
Summary of CInLPN object
}
