% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericFunctions.R
\name{taille}
\alias{taille}
\title{Function that compute the length of vector without NAs}
\usage{
taille(x)
}
\arguments{
\item{x}{numeric vector with possibly NAs}
}
\value{
a numeric vector
}
\description{
Function that compute the length of vector without NAs
}
