% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VecToMat}
\alias{VecToMat}
\title{Function that transforms a vector to a matrix}
\usage{
VecToMat(y, K, m_i)
}
\arguments{
\item{y}{a vector}

\item{K}{an integer indicating the number of columns of the returned matrix}

\item{m_i}{an integer indicating the number of rows of the returned matrix}
}
\value{
a matrix
}
\description{
Function that transforms a vector to a matrix
}
