% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaldMult.R
\name{WaldMult}
\alias{WaldMult}
\title{Multiple Wald test}
\usage{
WaldMult(v, beta, pos, matR, value)
}
\arguments{
\item{v}{vaiance-covariance matrix elements}

\item{beta}{estimates values}

\item{pos}{position of the parameters to test}

\item{matR}{indicates how to combine paerameters for the global test}

\item{value}{values of the right hand of test}
}
\value{
p-value
}
\description{
Multiple Wald test
}
