% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{YiNui}
\alias{YiNui}
\title{Function that computes the difference (mat_Yi - mat_Nu_i), delates missing values (NAs) and 
returns a vector. mat_Yi is the outcomes and mat_Nu_i is the expectation}
\usage{
YiNui(nD, matrixP, tau, tau_i, DeltaT, Yi, x0i, alpha_mu0, xi, alpha_mu,
  G_mat_A_0_to_tau_i)
}
\arguments{
\item{nD}{an integer indicating the number of processes}

\item{matrixP}{a matrix that matches markers to latent processes}

\item{tau}{a vector of integers indicating times}

\item{tau_i}{a vector of integers indicating times for individual i}

\item{DeltaT}{double that indicates the discretization step}

\item{Yi}{a matrix of the outcomes}

\item{x0i}{model.matrix for baseline's submodel}

\item{alpha_mu0}{a vector of parameters associated to the model.matrix for the baseline's submodel}

\item{xi}{model.matrix for change's submodel}

\item{alpha_mu}{a vector of parameters associated to the model.matrix for the change's submodel}

\item{G_mat_A_0_to_tau_i}{matrix containing  Prod(A_t)t=0,tau_i where A_t is the transition
matric containing at time t}
}
\value{
a vector
}
\description{
Function that computes the difference (mat_Yi - mat_Nu_i), delates missing values (NAs) and 
returns a vector. mat_Yi is the outcomes and mat_Nu_i is the expectation
}
