% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{YiwoNA}
\alias{YiwoNA}
\title{Function that returns Yi (a vector) without NAs values}
\usage{
YiwoNA(Yi)
}
\arguments{
\item{Yi}{a vector with possibly NAs}
}
\value{
a vector
}
\description{
Function that returns Yi (a vector) without NAs values
}
