% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{simulated data to run an exemple}
\format{A datasets of 3584 observations and 7 variables}
\usage{
data(data)
}
\description{
simulated data to run an exemple
}
\examples{
data(data)
head(data)
}
\keyword{datasets}
