% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.CInLPN.R
\name{fit.CInLPN}
\alias{fit.CInLPN}
\title{Marginal and subject-specific predictions for CInLPN objects}
\usage{
fit.CInLPN(object, newdata, MCnr = 10, ...)
}
\arguments{
\item{object}{CInLPN object}

\item{newdata}{dataset}

\item{MCnr}{an integer that gives the number of Monte Carlo iterations}

\item{\dots}{optional parameters}
}
\value{
list of marginal and subject-specific predictions
}
\description{
Marginal and subject-specific predictions for CInLPN objects
}
