% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CInLPN.R
\name{predict.CInLPN}
\alias{predict.CInLPN}
\title{Marginal predictions for CInLPN objects}
\usage{
\method{predict}{CInLPN}(object, newdata, TimeDiscretization = TRUE,
  MCnr = 10, ...)
}
\arguments{
\item{object}{CInLPN object}

\item{newdata}{dataset}

\item{MCnr}{an integer that gives the number of Monte Carlo iterations}

\item{\dots}{optional parameters}
}
\value{
list of marginal and subject-specific predictions
}
\description{
Marginal predictions for CInLPN objects
}
